config.py
from sqlmodel import Session, SQLModel, create_engine ,select
import os
from dotenv import load_dotenv

load_dotenv()

url = os.getenv("DATABASE_URL")

engine = create_engine(url, echo=True, pool_size=3, max_overflow=0)

def create_db_and_tables():
    SQLModel.metadata.create_all(engine)

def get_session():
    with Session(engine) as session:
        yield session
