pyproject.toml
[project]
name = "dependencyissue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "locust>=2.43.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "uuid7>=0.1.0",
    "uvicorn>=0.41.0",
]
